\name{facets-deprecated}
\alias{facets-deprecated}
\title{Deprecated functions in package \sQuote{facets}}

\description{
  These functions are provided for compatibility with older versions
  of \sQuote{facets} only, and will be defunct at the next release.
}

\details{

  The following functions are deprecated and will be made defunct; use
  the replacement indicated below:
  \itemize{

    \item{emcncf2: \code{\link{emcncf}}}

  }
}
