\name{plotSampleCNCF}
\alias{plotSampleCNCF}
\title{Plot the data and results for a single sample}
\description{
  Plots copy number log-ratio, variant allele log-odds ratio as well as
  fitted cellular fraction and integer copy numbers.
} 
\usage{
  plotSampleCNCF(jointseg, out, fit)
}
\arguments{
  \item{jointseg}{jointseg from the output of procSample}
  \item{out}{out from the output of procSample}
  \item{fit}{output of emcncf}
}
