\name{facetsY-package}
\alias{facetsY-package}
\alias{facetsY}
\docType{package}
\title{
  \packageTitle{facetsY}
}
\description{
  \packageDescription{facetsY}
}
\details{
  FACETS Y stands for Fraction and Allelic Copy number Estimation from
  Tumor/normal Sequencing. INCLUDING Y CHROMOSOME. It uses high coverage next generation
  sequencing data (whole exome or targeted) to estimate the total and
  integer allele specific copy numbers as well as the tumor purity and
  ploidy. The methodology is described in the reference paper.
}
\author{
\packageAuthor{facetsY}

Maintainer: \packageMaintainer{facetsY}
}
\references{
  Shen, R and Seshan, VE (2016). FACETS: allele-specific copy number and
  clonal heterogeneity analysis tool for high-throughput DNA sequencing.
  \emph{Nucleic Acids Research} \url{https://doi.org/10.1093/nar/gkw520} 
}
\keyword{package}
